requirements.txt
requests
git+https://github.com/Pycord-Development/pycord
python-dotenv
aiocron